+++ types/index.d.ts
@@ -190,42 +190,55 @@ declare module "components/Button/index" {
     import PropTypes from "prop-types";
 }
 declare module "tokens/forimport/index.es" {
+    export var brandBlueXlight: string;
     export var brandBlueLight: string;
     export var brandBlueBase: string;
     export var brandBlueDark: string;
+    export var brandPurpleXlight: string;
     export var brandPurpleLight: string;
     export var brandPurpleBase: string;
     export var brandPurpleDark: string;
+    export var smokeXlight: string;
     export var smokeLight: string;
     export var smokeBase: string;
     export var smokeDark: string;
+    export var greyXlight: string;
     export var greyLight: string;
     export var greyBase: string;
     export var greyDark: string;
+    export var purpleXlight: string;
     export var purpleLight: string;
     export var purpleBase: string;
     export var purpleDark: string;
+    export var magentaXlight: string;
     export var magentaLight: string;
     export var magentaBase: string;
     export var magentaDark: string;
-    export var redLight: string;
-    export var redBase: string;
-    export var redDark: string;
+    export var pinkXlight: string;
     export var pinkLight: string;
     export var pinkBase: string;
     export var pinkDark: string;
+    export var redXlight: string;
+    export var redLight: string;
+    export var redBase: string;
+    export var redDark: string;
+    export var orangeXlight: string;
     export var orangeLight: string;
     export var orangeBase: string;
     export var orangeDark: string;
-    export var amberLight: string;
-    export var amberBase: string;
-    export var amberDark: string;
+    export var yellowXlight: string;
+    export var yellowLight: string;
+    export var yellowBase: string;
+    export var yellowDark: string;
+    export var greenXlight: string;
     export var greenLight: string;
     export var greenBase: string;
     export var greenDark: string;
+    export var aquaXlight: string;
     export var aquaLight: string;
     export var aquaBase: string;
     export var aquaDark: string;
+    export var blueXlight: string;
     export var blueLight: string;
     export var blueBase: string;
     export var blueDark: string;
@@ -640,7 +653,7 @@ declare module "components/ButtonRow/index" {
 }
 declare module "components/Link/index" {
     export default Link;
-    function Link({ ariaLabel, children, href, onClick, isDisabled, isFullWidth, newWindow, style, testSection, }: any): any;
+    function Link({ ariaLabel, children, href, leftIcon, onClick, isDisabled, isFullWidth, newWindow, rightIcon, style, testSection, }: any): any;
     namespace Link {
         export namespace propTypes {
             export const ariaLabel: PropTypes.Requireable<string>;
@@ -648,8 +661,10 @@ declare module "components/Link/index" {
             export const href: PropTypes.Requireable<string>;
             export const isDisabled: PropTypes.Requireable<boolean>;
             export const isFullWidth: PropTypes.Requireable<boolean>;
+            export const leftIcon: PropTypes.Requireable<PropTypes.ReactNodeLike>;
             export const newWindow: PropTypes.Requireable<boolean>;
             export const onClick: PropTypes.Requireable<(...args: any[]) => any>;
+            export const rightIcon: PropTypes.Requireable<PropTypes.ReactNodeLike>;
             export const style: PropTypes.Requireable<string>;
             export const testSection: PropTypes.Requireable<string>;
         }
@@ -1137,10 +1152,12 @@ declare module "components/DatePicker/DatePicker" {
         onOutsideClick: () => void;
         onClose: () => void;
         isOutsideAcceptableDateRange: (day: any) => any;
+        onClearButtonClick: () => void;
         render(): JSX.Element;
     }
     namespace DatePicker {
         export namespace propTypes {
+            export const hasClearButton: PropTypes.Requireable<boolean>;
             export const initialDate: any;
             export const initialVisibleMonth: PropTypes.Requireable<(...args: any[]) => any>;
             export const inputId: PropTypes.Validator<string>;
@@ -1230,6 +1247,7 @@ declare module "components/DatePicker/DateRangePicker" {
         isOutsideAcceptableDateRange: (day: any) => any;
         renderPresetButtons: (presetPanelOptions: any) => any;
         renderSidePanel: () => JSX.Element;
+        onClearButtonClick: (rangeSide: any) => void;
         render(): JSX.Element;
     }
     namespace DateRangePicker {
@@ -1240,6 +1258,7 @@ declare module "components/DatePicker/DateRangePicker" {
             export { timeInputValidator as endTimeInputId };
             export const endTimeInputLabel: PropTypes.Requireable<string>;
             export const focusedInput: PropTypes.Requireable<string | null>;
+            export const hasClearButton: PropTypes.Requireable<boolean>;
             export const hasTimeInputs: PropTypes.Requireable<boolean>;
             export const initialEndDate: any;
             export const initialStartDate: any;
@@ -1578,7 +1597,7 @@ declare module "components/Dropdown/DropdownBlockLinkText/index" {
             export const hasExternalIcon: PropTypes.Requireable<boolean>;
             export const isDestructive: PropTypes.Requireable<boolean>;
             export const testSection: PropTypes.Requireable<string>;
-            const text_1: PropTypes.Requireable<string>;
+            const text_1: PropTypes.Requireable<PropTypes.ReactNodeLike>;
             export { text_1 as text };
         }
     }
@@ -1595,7 +1614,7 @@ declare module "components/Dropdown/DropdownBlockLinkSecondaryText/index" {
         export namespace propTypes {
             const isWarning_1: PropTypes.Requireable<boolean>;
             export { isWarning_1 as isWarning };
-            const secondaryText_1: PropTypes.Requireable<string>;
+            const secondaryText_1: PropTypes.Requireable<PropTypes.ReactNodeLike>;
             export { secondaryText_1 as secondaryText };
             export const testSection: PropTypes.Requireable<string>;
         }